Template matching code now iterates over all objects/scales/rotations

+++ helpers.py
@@ -1,6 +1,3 @@
-from os import listdir
-
-
 def white_to_black(img):
     for m in range(0, img.shape[0]):
         for n in range(0, img.shape[1]):
@@ -13,8 +10,8 @@ def get_object_name(file):
     return file[4:len(file) - 4]
 
 
-def get_scale_percentage(depth, n=100):
+def get_scale_percentage(depth, n=100.0):
     if depth == 0:
         return n
 
-    return get_scale_percentage(depth - 1, int(n / 2))
+    return get_scale_percentage(depth - 1, n / 2)
